+++ inst/tinytest/bind_array_1_basic/test-bind_array-errors.R
@@ -231,3 +231,4 @@ expect_silent(
 )
 enumerate <- enumerate + 8L
 
+